Added CSS folder, data folder and moved files to correct spots
